refactor(ApplicationRegister): Member를 도메인 이벤트를 적용하여 생성하도록 변경
