+++ src/main/java/page/clab/api/domain/hiring/application/application/event/ApplicationEventProcessor.java
@@ -1,0 +1,61 @@
+package page.clab.api.domain.hiring.application.application.event;
+
+import jakarta.annotation.PostConstruct;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Component;
+import page.clab.api.domain.hiring.application.domain.Application;
+import page.clab.api.domain.memberManagement.member.application.event.MemberEventProcessor;
+import page.clab.api.domain.memberManagement.member.application.event.MemberEventProcessorRegistry;
+import page.clab.api.domain.memberManagement.member.application.port.out.RegisterMemberPort;
+import page.clab.api.domain.memberManagement.member.domain.Member;
+import page.clab.api.global.common.email.application.EmailService;
+import page.clab.api.global.common.verification.application.VerificationService;
+
+import java.util.concurrent.CompletableFuture;
+
+@Slf4j
+@Component
+@RequiredArgsConstructor
+public class ApplicationEventProcessor implements MemberEventProcessor {
+    private final VerificationService verificationService;
+    private final EmailService emailService;
+    private final PasswordEncoder passwordEncoder;
+    private final RegisterMemberPort registerMemberPort;
+    private final MemberEventProcessorRegistry processorRegistry;
+
+    @PostConstruct
+    public void init() {
+        processorRegistry.register(this);
+    }
+
+    @Override
+    public void processMemberDeleted(String memberId) {
+        // do nothing
+    }
+
+    @Override
+    public void processMemberUpdated(String memberId) {
+        // do nothing
+    }
+
+    @Override
+    public void processMemberCreated(Application application) {
+        Member member = Member.fromApplication(application);
+        setRandomPasswordAndSendEmail(member);
+        registerMemberPort.save(member);
+    }
+
+    private void setRandomPasswordAndSendEmail(Member member) {
+        String password = verificationService.generateVerificationCode();
+        member.updatePassword(password, passwordEncoder);
+        CompletableFuture.runAsync(() -> {
+            try {
+                emailService.broadcastEmailToApprovedMember(member, password);
+            } catch (Exception e) {
+                log.error("이메일 전송 실패: {}", e.getMessage());
+            }
+        });
+    }
+}
