+++ src/main/java/page/clab/api/domain/hiring/application/application/service/ApplicationMemberRegisterService.java
@@ -2,26 +2,23 @@ package page.clab.api.domain.hiring.application.application.service;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import page.clab.api.domain.hiring.application.application.exception.NotApprovedApplicationException;
 import page.clab.api.domain.hiring.application.application.port.in.RegisterMembersByRecruitmentUseCase;
 import page.clab.api.domain.hiring.application.application.port.out.RetrieveApplicationPort;
 import page.clab.api.domain.hiring.application.domain.Application;
+import page.clab.api.domain.memberManagement.member.application.event.MemberCreatedEvent;
 import page.clab.api.domain.memberManagement.member.domain.Member;
 import page.clab.api.domain.memberManagement.position.domain.Position;
 import page.clab.api.domain.memberManagement.position.domain.PositionType;
-import page.clab.api.external.memberManagement.member.application.port.ExternalRegisterMemberUseCase;
 import page.clab.api.external.memberManagement.member.application.port.ExternalRetrieveMemberUseCase;
 import page.clab.api.external.memberManagement.position.application.port.ExternalRegisterPositionUseCase;
 import page.clab.api.external.memberManagement.position.application.port.ExternalRetrievePositionUseCase;
-import page.clab.api.global.common.email.application.EmailService;
-import page.clab.api.global.common.verification.application.VerificationService;
 
 import java.time.LocalDate;
 import java.util.List;
-import java.util.concurrent.CompletableFuture;
 
 @Service
 @RequiredArgsConstructor
@@ -29,13 +26,10 @@ import java.util.concurrent.CompletableFuture;
 public class ApplicationMemberRegisterService implements RegisterMembersByRecruitmentUseCase {
 
     private final RetrieveApplicationPort retrieveApplicationPort;
-    private final ExternalRegisterMemberUseCase externalRegisterMemberUseCase;
     private final ExternalRetrieveMemberUseCase externalRetrieveMemberUseCase;
     private final ExternalRegisterPositionUseCase externalRegisterPositionUseCase;
     private final ExternalRetrievePositionUseCase externalRetrievePositionUseCase;
-    private final EmailService emailService;
-    private final VerificationService verificationService;
-    private final PasswordEncoder passwordEncoder;
+    private final ApplicationEventPublisher eventPublisher;
 
     @Transactional
     @Override
@@ -69,23 +63,10 @@ public class ApplicationMemberRegisterService implements RegisterMembersByRecrui
     private Member createMemberByApplication(Application application) {
         return externalRetrieveMemberUseCase.findById(application.getStudentId())
                 .orElseGet(() -> {
-                    Member member = Application.toMember(application);
-                    setRandomPasswordAndSendEmail(member);
-                    externalRegisterMemberUseCase.save(member);
+                    Member member = Member.fromApplication(application);
+                    eventPublisher.publishEvent(new MemberCreatedEvent(this, application));
                     return member;
                 });
-    }
-
-    private void setRandomPasswordAndSendEmail(Member member) {
-        String password = verificationService.generateVerificationCode();
-        member.updatePassword(password, passwordEncoder);
-        CompletableFuture.runAsync(() -> {
-            try {
-                emailService.broadcastEmailToApprovedMember(member, password);
-            } catch (Exception e) {
-                log.error("이메일 전송 실패: {}", e.getMessage());
-            }
-        });
     }
 
     public void createPositionByMember(Member member) {
