+++ src/main/java/page/clab/api/domain/memberManagement/member/application/event/MemberEventProcessor.java
@@ -1,8 +1,12 @@
 package page.clab.api.domain.memberManagement.member.application.event;
 
+import page.clab.api.domain.hiring.application.domain.Application;
+
 public interface MemberEventProcessor {
 
     void processMemberDeleted(String memberId);
 
     void processMemberUpdated(String memberId);
+
+    void processMemberCreated(Application application);
 }
